refactor(lib/table): migrar lógica de directiva a componente y mejorar soporte de temas
