+++ projects/ngx-data-visualizer/src/lib/table/table.component.scss
@@ -20,37 +20,3 @@ $z-index-inner: 9;
   }
 
 }
-
-/**
- * Botón para abrir/cerrar el editor de configuración
- */
-.lib-table-config-toggle {
-  position: absolute;
-  top: 8px;
-  right: 8px;
-  z-index: $z-index-inner + 1;
-  background: rgba(255, 255, 255, 0.8);
-  backdrop-filter: blur(4px);
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-  width: 36px;
-  height: 36px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  color: #4a5568;
-  transition: all 0.2s;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-
-  &:hover {
-    background: #fff;
-    color: #2b6cb0;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    transform: translateY(-1px);
-  }
-
-  &:active {
-    transform: translateY(0);
-  }
-}
